Fixed the Travis CI builds.
## The Problem
Travis CI is failing all of our builds because they require a NeverBounce API
key, which is stored in a secure Environment Variable.

Unfortunately, Travis CI prevents external forks from accessing this variable.

See
 - https://docs.travis-ci.com/user/environment-variables/
 - https://github.com/google/containerregistry/issues/114

## The Solution
We're going to follow the practices of the OHDSI / Achilles Project:
 - https://github.com/OHDSI/Achilles/issues/165

0. Only unit tests will run for third-party repos, including our
company's developer's forks (unfortunately!).
1. After code review, the build master shall push the developer's fork
to the mainline PHP Experts' repository and wait for the full tests to
run.
2. After the Pull Request is merged, the build master shall delete the
mainline branch, but keep the contributor's branch, per our existing
policy.

For #5

# Please enter the commit message for your changes. Lines starting
# with '#' will be kept; you may remove them yourself if you want to.
# An empty message aborts the commit.
#
# Date:      Wed Jun 5 07:32:37 2019 -0500
#
# On branch 5-fix_travis_ci
# Changes to be committed:
#	modified:   .travis.yml
#
# Untracked files:
#	build/
#	composer-with_phpunit.log
#	composer-without_phpunit.log
#	env
#	travis
#